Split background.js file into ES6 Modules
