+++ background/background.js
@@ -1,0 +1,10 @@
+"use strict";
+
+import * as IconHandler from "./modules/IconHandler.js";
+import * as ContextMenu from "./modules/ContextMenu.js";
+import * as BrowserCommunication from "./modules/BrowserCommunication.js";
+
+// init modules
+IconHandler.init();
+ContextMenu.init();
+BrowserCommunication.init();
